# Use variadic args in constructor