--- a/src/DI/Definition/Resolver/ObjectCreator.php
+++ b/src/DI/Definition/Resolver/ObjectCreator.php
@@ -146,27 +146,7 @@ class ObjectCreator implements DefinitionResolver
                 $parameters
             );
 
-            // Optimization trick
-            switch (count($args)) {
-                case 0:
-                    $object = new $classname;
-                    break;
-                case 1:
-                    $object = new $classname($args[0]);
-                    break;
-                case 2:
-                    $object = new $classname($args[0], $args[1]);
-                    break;
-                case 3:
-                    $object = new $classname($args[0], $args[1], $args[2]);
-                    break;
-                case 4:
-                    $object = new $classname($args[0], $args[1], $args[2], $args[3]);
-                    break;
-                default:
-                    $object = $classReflection->newInstanceArgs($args);
-                    break;
-            }
+            $object = new $classname(...$args);
 
             $this->injectMethodsAndProperties($object, $definition);
         } catch (NotFoundException $e) {