Though this could be optimised without the extra verbosity. Crude benchmarks below:

### Reflection
Code
```php
class Foo { public function __construct($a, $b, $c, $d) {}  }
$foo = Foo::class;
$args = [1,2,3,4];
$reflect = new ReflectionClass($foo);
for ($i = 0; $i < 10000000; $i++) {
    $reflect->newInstanceArgs($args);
}
```
Time
```bash
$ time php reflect.php 
real	0m 3.42s
user	0m 3.41s
sys	0m 0.00s
```

### Current
Code
```php
class Foo { public function __construct($a, $b, $c, $d) {}  }
$foo = Foo::class;
$args = [1,2,3,4];
for ($i = 0; $i < 10000000; $i++) {
    new $foo($args[0], $args[1], $args[2], $args[3]);
}
```
Time
```bash
$ time php current.php 
real	0m 2.85s
user	0m 2.84s
sys	0m 0.00s
```

### Variadic
Code
```php
class Foo { public function __construct($a, $b, $c, $d) {}  }
$foo = Foo::class;
$args = [1,2,3,4];
for ($i = 0; $i < 10000000; $i++) {
    new $foo(...$args);
}
```
Time
```bash
$ time php variadic.php 
real	0m 2.48s
user	0m 2.48s
sys	0m 0.00s
```